# docs/swagger.yaml
definitions:
...
  dao.TaskInfo:
    properties:
      auth_scan:
        type: integer
      create_at:
        type: string
      id:
        type: integer
      is_delete:
        type: integer
      name:
        type: string
      percent:
        type: integer
      poc_scan:
        type: integer
      progress:
        type: string
      recursion:
        type: integer
      rule_id:
        type: integer
      status:
        type: string
      target_list:
        type: string
      trap_scan:
        type: integer
      update_at:
        type: string
      web_scan:
        type: integer
    type: object
...
  dto.ChartBoxCard:
    properties:
      image:
        description: 卡片头像
        type: string
      series:
        description: 数据源
        items:
          $ref: '#/definitions/dto.ChartSeries'
        type: array
      title:
        description: 卡片标题
        type: string
      type:
        description: 卡片绘图类型 pie line tatle
        type: string
    type: object
  dto.ChartSeries:
    properties:
      name:
        type: string
      value:
        type: integer
    type: object
...
  dto.PanelGroupData:
    properties:
      asset_count:
        type: integer
      resource_count:
        type: integer
      service_count:
        type: integer
      vul_count:
        type: integer
    type: object
...
  dto.TableBoxCard:
    properties:
      image:
        description: 卡片头像
        type: string
      series:
        description: 数据源
        items:
          $ref: '#/definitions/dto.VulInfoOutput'
        type: array
      title:
        description: 卡片标题
        type: string
      type:
        description: 卡片绘图类型 pie line tatle
        type: string
    type: object
  dto.TaskAddInput:
    properties:
      auth_scan:
        type: integer
      name:
        type: string
      poc_scan:
        type: integer
      recursion:
        type: integer
      rule_id:
        type: integer
      target_list:
        type: string
      trap_scan:
        type: integer
      web_scan:
        type: integer
    required:
    - name
    - target_list
    type: object
  dto.TaskDashboardOutput:
    properties:
      box1:
        $ref: '#/definitions/dto.ChartBoxCard'
        type: object
      box2:
        $ref: '#/definitions/dto.ChartBoxCard'
        type: object
      box3:
        $ref: '#/definitions/dto.ChartBoxCard'
        type: object
      box4:
        $ref: '#/definitions/dto.ChartBoxCard'
        type: object
      box5:
        $ref: '#/definitions/dto.TableBoxCard'
        type: object
      box6:
        $ref: '#/definitions/dto.ChartBoxCard'
        type: object
      panel_data:
        $ref: '#/definitions/dto.PanelGroupData'
        type: object
      percent:
        type: integer
      status:
        type: string
    type: object
  dto.TaskInfoOutput:
    properties:
      asset_num:
        type: integer
      auth_scan:
        type: integer
      create_at:
        type: string
      id:
        type: integer
      is_delete:
        type: integer
      name:
        type: string
      percent:
        type: integer
      poc_scan:
        type: integer
      progress:
        type: string
      recursion:
        type: integer
      rule_id:
        type: integer
      status:
        type: string
      target_list:
        type: string
      trap_scan:
        type: integer
      update_at:
        type: string
      web_scan:
        type: integer
    type: object
  dto.TaskListOutput:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.TaskInfoOutput'
        type: array
      total:
        type: integer
    type: object
  dto.TaskUpdateInput:
    properties:
      auth_scan:
        type: integer
      id:
        type: integer
      name:
        type: string
      poc_scan:
        type: integer
      recursion:
        type: integer
      rule_id:
        type: integer
      target_list:
        type: string
      trap_scan:
        type: integer
      web_scan:
        type: integer
    required:
    - id
    - name
    - target_list
    type: object
...
  dto.TrapPluginAddInput:
    properties:
      app_name:
        type: string
      author:
        type: string
      content:
        type: string
      desc:
        type: string
      honeypot:
        type: string
      name:
        type: string
      protocol:
        type: string
      trap_id:
        type: string
    required:
    - app_name
    - author
    - content
    - desc
    - honeypot
    - name
    - protocol
    - trap_id
    type: object
  dto.TrapPluginListOutput:
    properties:
      list:
        description: 列表
        items:
          $ref: '#/definitions/dto.TrapPluginOutput'
        type: array
      total:
        description: 总数
        type: integer
    type: object
  dto.TrapPluginOutput:
    properties:
      app_name:
        type: string
      author:
        type: string
      content:
        type: string
      create_at:
        type: string
      desc:
        type: string
      honeypot:
        type: string
      id:
        type: integer
      is_delete:
        type: integer
      name:
        type: string
      protocol:
        type: string
      trap_id:
        type: string
      update_at:
        type: string
    type: object
  dto.TrapPluginUpdateInput:
    properties:
      app_name:
        type: string
      author:
        type: string
      content:
        type: string
      desc:
        type: string
      honeypot:
        type: string
      id:
        type: integer
      name:
        type: string
      protocol:
        type: string
      trap_id:
        type: string
    required:
    - app_name
    - author
    - content
    - desc
    - honeypot
    - id
    - name
    - protocol
    - trap_id
    type: object
...
  dto.VulInfoOutput:
    properties:
      app_name:
        type: string
      asset:
        type: string
      asset_id:
        type: integer
      create_at:
        type: string
      exploit_payload:
        type: string
      exploit_result:
        type: string
      exploit_url:
        type: string
      id:
        type: integer
      is_delete:
        type: integer
      line:
        type: integer
      plugin_id:
        type: integer
      port_id:
        type: integer
      trojan_payload:
        type: string
      trojan_result:
        type: string
      trojan_url:
        type: string
      verify_payload:
        type: string
      verify_result:
        type: string
      verify_url:
        type: string
      vul_name:
        type: string
      vul_type:
        type: string
      webshell_payload:
        type: string
      webshell_result:
        type: string
      webshell_url:
        type: string
    type: object
...
  middleware.Response:
    properties:
      data:
        type: object
      errmsg:
        type: string
      errno:
        type: integer
      stack:
        type: object
      trace_id:
        type: object
    type: object
info:
  contact: {}
  license: {}
paths:
...
  /task/add:
    post:
      consumes:
      - application/json
      description: 任务添加
      operationId: /task/add
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TaskAddInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 任务添加
      tags:
      - 任务管理
  /task/delete:
    delete:
      consumes:
      - application/json
      description: 任务删除
      operationId: /task/delete
      parameters:
      - description: 任务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 任务删除
      tags:
      - 任务管理
  /task/detail:
    get:
      consumes:
      - application/json
      description: 任务详情(静态视图)
      operationId: /task/detail
      parameters:
      - description: 任务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskDashboardOutput'
              type: object
      summary: 任务详情(静态视图)
      tags:
      - 任务管理
  /task/exec:
    patch:
      consumes:
      - application/json
      description: 执行任务
      operationId: /task/exec
      parameters:
      - description: 任务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 执行任务
      tags:
      - 任务管理
  /task/info:
    get:
      consumes:
      - application/json
      description: 任务详情
      operationId: /task/info
      parameters:
      - description: 任务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dao.TaskInfo'
              type: object
      summary: 任务详情
      tags:
      - 任务管理
  /task/list:
    get:
      consumes:
      - application/json
      description: 任务列表
      operationId: /task/list
      parameters:
      - description: 模糊查询
        in: query
        name: info
        type: string
      - description: 每页条数
        in: query
        name: limit
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskListOutput'
              type: object
      summary: 任务列表
      tags:
      - 任务管理
  /task/stat:
    get:
      consumes:
      - application/json
      description: 任务视图
      operationId: /task/stat/get
      parameters:
      - description: 任务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskDashboardOutput'
              type: object
      summary: 任务视图
      tags:
      - 任务管理
  /task/update:
    put:
      consumes:
      - application/json
      description: 任务更新
      operationId: /task/update
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TaskUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 任务更新
      tags:
      - 任务管理
  /trap/plugin/add:
    post:
      consumes:
      - application/json
      description: 插件添加
      operationId: /trap/plugin/add
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TrapPluginAddInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 插件添加
      tags:
      - 蜜罐识别
  /trap/plugin/delete:
    delete:
      consumes:
      - application/json
      description: 插件删除
      operationId: /trap/plugin/delete
      parameters:
      - description: 任务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 插件删除
      tags:
      - 蜜罐识别
  /trap/plugin/info:
    get:
      consumes:
      - application/json
      description: 插件信息
      operationId: /trap/plugin/info
      parameters:
      - description: 任务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dao.TaskInfo'
              type: object
      summary: 插件信息
      tags:
      - 蜜罐识别
  /trap/plugin/list:
    get:
      consumes:
      - application/json
      description: 插件列表
      operationId: /trap/plugin/list
      parameters:
      - description: 模糊查询
        in: query
        name: info
        type: string
      - description: 每页条数
        in: query
        name: limit
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TrapPluginListOutput'
              type: object
      summary: 插件列表
      tags:
      - 蜜罐识别
  /trap/plugin/update:
    put:
      consumes:
      - application/json
      description: 插件更新
      operationId: /trap/plugin/update
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TrapPluginUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 插件更新
      tags:
      - 蜜罐识别
...
swagger: "2.0"